moving confirmations test into integrationTest for simplicity
